Things are starting to draw
I can draw a triangle, but I need to get textures to work now. Tomorrow
will be spent trying to get to that point. Once i've got textures going
here then i'll be able to start working on animation and movement
working. From there it'll be down to getting text working and then
starting to build a game.